Behind-the-scenes changes on faradayScan and Rotation. Making things a bit more modular where one simply makes a call to "get fourier data" to make the motor rotate and monitor the diodes. This also makes it more flexible in that you can change the photodiodes that are being monitored or the motors which are being rotated in one place. In the grand scheme of things, this should actually be used by the polarimeter too, I think. The current version of faraday rotation also has a feature where you specify the step size and the number of rotations. I'm doing a short study into finding the best values to use at which point I'll send it back to its original form and fix the step size and number of rotations.
This is where the stepSizeStudy.sh script comes in. It's the script I
run to collect the data.

FaradayScanAnalysisTools then needed an update so that it could process
the data in the new way that faradayScan and Rotation record it.
Overall, I think it's a lot of good changes that make a lot more sense
programmatically.